[4] improve health check documentation
