+++ src/routes/mod.rs
@@ -1,12 +1,12 @@
 use axum::Router;
 
-pub mod health;
+pub mod health_check;
 pub mod swagger;
 pub mod teapot;
 
 pub fn routes() -> Router {
     Router::new()
-        .merge(health::route())
+        .merge(health_check::route())
         .merge(teapot::route())
         .merge(swagger::route())
 }
