+++ src/routes/swagger.rs
@@ -2,13 +2,14 @@ use axum::Router;
 use utoipa::OpenApi;
 use utoipa_swagger_ui::SwaggerUi;
 
-use super::health;
+use super::health_check;
 use super::teapot;
 
 pub fn route() -> Router {
-    Router::new().merge(SwaggerUi::new("/swagger-ui").url("/openapi.json", ApiDoc::openapi()))
+    Router::new()
+        .merge(SwaggerUi::new("/swagger-ui").url("/openapi.json", ApiDoc::openapi()))
 }
 
 #[derive(OpenApi)]
-#[openapi(paths(health::route, teapot::route))]
+#[openapi(paths(health_check::live, health_check::health, teapot::route))]
 pub struct ApiDoc;
